requirements.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
ecdsa==0.14.1
future==0.18.2
gunicorn==20.0.0
idna==2.8
psycopg2-binary==2.8.4
pyasn1==0.4.8
PyJWT==1.7.1
python-decouple==3.3
python-jose==3.0.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==4.1.4
